refactor(desktop): consolidate activity render class presenters
- Move transcript item presentation for the 12 activity render classes into sibling components under activityRenderClasses.
- Add a typed render-class presenter registry so every AgentActivityRenderClass maps to a presentational component.
- Keep AgentSessionTranscriptList focused on transcript display blocks, turn grouping, setup rows, and prompt context rendering.
- Share timestamp, identity props, lifecycle, and tool row delegates across the extracted presenters without changing transcript behavior.
