+++ desktop/src/features/agents/ui/AgentSessionTranscriptList.tsx
@@ -1,25 +1,19 @@
 import * as React from "react";
-import {
-  AlertCircle,
-  Brain,
-  CheckCheck,
-  ChevronDown,
-  CircleDot,
-  Radio,
-  TerminalSquare,
-} from "lucide-react";
+import { CheckCheck, ChevronDown, Radio } from "lucide-react";
 
 import {
   resolveUserLabel,
   type UserProfileLookup,
 } from "@/features/profile/lib/identity";
 import { cn } from "@/shared/lib/cn";
-import { normalizePubkey } from "@/shared/lib/pubkey";
 import { Markdown } from "@/shared/ui/markdown";
 import { Toggle } from "@/shared/ui/toggle";
 import { UserAvatar } from "@/shared/ui/UserAvatar";
+import { Tooltip, TooltipContent, TooltipTrigger } from "@/shared/ui/tooltip";
 import type { PromptSection, TranscriptItem } from "./agentSessionTypes";
-import { ToolItem } from "./AgentSessionToolItem";
+import { TranscriptActivityItem } from "./activityRenderClasses/TranscriptActivityItem";
+import { TranscriptTimestamp } from "./activityRenderClasses/TranscriptTimestamp";
+import type { AgentTranscriptIdentityProps } from "./activityRenderClasses/types";
 import {
   buildTranscriptDisplayBlocks,
   formatTurnSetupLabel,
@@ -28,8 +22,6 @@ import {
   type TranscriptDisplayBlock,
   type TranscriptTurnSegment,
 } from "./agentSessionTranscriptGrouping";
-import { formatTranscriptTime } from "./agentSessionUtils";
-import { Tooltip, TooltipContent, TooltipTrigger } from "@/shared/ui/tooltip";
 
 const TRANSCRIPT_ACP_SOURCE_STORAGE_KEY = "buzz:show-transcript-acp-source";
 
@@ -57,12 +49,6 @@ function shouldShowTranscriptAcpSource() {
     return false;
   }
 }
-
-type AgentTranscriptIdentityProps = {
-  agentAvatarUrl: string | null;
-  agentName: string;
-  agentPubkey: string;
-};
 
 export function AgentSessionTranscriptList({
   agentAvatarUrl,
@@ -542,264 +528,13 @@ const TranscriptItemView = React.memo(function TranscriptItemView({
   item: TranscriptItem;
   profiles?: UserProfileLookup;
 }) {
-  if (item.type === "message") {
-    return (
-      <MessageItem
-        agentAvatarUrl={agentAvatarUrl}
-        agentName={agentName}
-        agentPubkey={agentPubkey}
-        item={item}
-        profiles={profiles}
-      />
-    );
-  }
-  if (item.type === "tool") {
-    return <ToolItem item={item} />;
-  }
-  if (item.type === "thought") {
-    return <ThoughtItem item={item} />;
-  }
-  if (item.type === "plan") {
-    return <PlanItem item={item} />;
-  }
-  if (item.type === "metadata") {
-    return <MetadataItem item={item} />;
-  }
-  return <LifecycleItem item={item} />;
+  return (
+    <TranscriptActivityItem
+      agentAvatarUrl={agentAvatarUrl}
+      agentName={agentName}
+      agentPubkey={agentPubkey}
+      item={item}
+      profiles={profiles}
+    />
+  );
 });
-
-function MessageItem({
-  agentAvatarUrl,
-  agentName,
-  agentPubkey,
-  item,
-  profiles,
-}: AgentTranscriptIdentityProps & {
-  item: Extract<TranscriptItem, { type: "message" }>;
-  profiles?: UserProfileLookup;
-}) {
-  const isAssistant = item.role === "assistant";
-  const text = item.text.trim();
-  const authorProfile = item.authorPubkey
-    ? profiles?.[item.authorPubkey.toLowerCase()]
-    : null;
-  const authorLabel = item.authorPubkey
-    ? resolveUserLabel({
-        pubkey: item.authorPubkey,
-        fallbackName: item.title,
-        profiles,
-      })
-    : item.title || "User";
-  const agentProfile = profiles?.[normalizePubkey(agentPubkey)] ?? null;
-  const assistantLabel = resolveUserLabel({
-    pubkey: agentPubkey,
-    fallbackName: agentName,
-    profiles,
-    preferResolvedSelfLabel: true,
-  });
-  const assistantAvatarUrl = agentProfile?.avatarUrl ?? agentAvatarUrl;
-
-  return (
-    <div
-      className={cn(
-        "flex animate-in fade-in duration-200 motion-reduce:animate-none",
-        isAssistant
-          ? "flex-row px-0 py-1.5"
-          : "flex-row items-start justify-end px-0 py-0.5",
-      )}
-      data-role={isAssistant ? "assistant-message" : "user-message"}
-      data-testid={
-        isAssistant ? "transcript-assistant-message" : "transcript-user-message"
-      }
-    >
-      {!isAssistant ? (
-        <UserAvatar
-          avatarUrl={authorProfile?.avatarUrl ?? null}
-          className="order-last ml-2 mt-1 shrink-0"
-          displayName={authorLabel}
-          size="xs"
-        />
-      ) : null}
-      <div
-        className={cn(
-          "group relative flex min-w-0 flex-col gap-1",
-          isAssistant ? "w-full items-start" : "max-w-[85%] items-end",
-        )}
-      >
-        {isAssistant ? (
-          <div className="mb-0.5 flex items-center gap-1.5 text-xs">
-            <UserAvatar
-              avatarUrl={assistantAvatarUrl}
-              className="shrink-0"
-              displayName={assistantLabel}
-              size="xs"
-              testId="transcript-assistant-avatar"
-            />
-            <span className="text-xs font-semibold text-foreground">
-              {assistantLabel}
-            </span>
-            <TranscriptTimestamp timestamp={item.timestamp} />
-          </div>
-        ) : null}
-        <div
-          className={cn(
-            "w-full min-w-0 text-sm leading-relaxed",
-            !isAssistant && "rounded-2xl bg-muted p-3 text-foreground",
-          )}
-        >
-          {isAssistant ? (
-            <Markdown compact content={text || " "} />
-          ) : (
-            <>
-              <Markdown content={text || " "} mediaInset tight />
-              <TranscriptTimestamp timestamp={item.timestamp} />
-            </>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function ThoughtItem({
-  item,
-}: {
-  item: Extract<TranscriptItem, { type: "thought" }>;
-}) {
-  return (
-    <details
-      className="group not-prose w-full rounded-md border border-transparent px-0"
-      data-testid="transcript-thought-item"
-    >
-      <summary className="inline-flex max-w-full cursor-pointer list-none items-center gap-1.5 py-px text-muted-foreground">
-        <Brain className="h-4 w-4" />
-        <span className="truncate text-sm font-medium">{item.title}</span>
-        <TranscriptTimestamp timestamp={item.timestamp} />
-        <ChevronDown className="h-4 w-4 shrink-0 transition-transform group-open:rotate-180" />
-      </summary>
-      <div className="py-2 pl-5 text-sm leading-6 text-muted-foreground">
-        <Markdown compact content={item.text.trim() || " "} />
-      </div>
-    </details>
-  );
-}
-
-function PlanItem({
-  item,
-}: {
-  item: Extract<TranscriptItem, { type: "plan" }>;
-}) {
-  return (
-    <details
-      className="group not-prose w-full rounded-md border border-primary/15 bg-primary/5 px-2 py-1"
-      data-testid="transcript-plan-item"
-      open
-    >
-      <summary className="inline-flex max-w-full cursor-pointer list-none items-center gap-1.5 py-px text-primary/90">
-        <CheckCheck className="h-3.5 w-3.5 shrink-0" />
-        <span className="truncate text-xs font-medium">{item.title}</span>
-        <TranscriptTimestamp timestamp={item.timestamp} />
-        <ChevronDown className="h-4 w-4 shrink-0 transition-transform group-open:rotate-180" />
-      </summary>
-      <div className="py-2 pl-5 text-sm leading-6 text-muted-foreground">
-        <Markdown compact content={item.text.trim() || "No plan details."} />
-      </div>
-    </details>
-  );
-}
-
-function MetadataItem({
-  item,
-}: {
-  item: Extract<TranscriptItem, { type: "metadata" }>;
-}) {
-  return (
-    <details
-      className="group not-prose w-full rounded-md border border-border/50 bg-muted/20 px-2 py-1"
-      data-testid="transcript-metadata-item"
-    >
-      <summary className="inline-flex max-w-full cursor-pointer list-none items-center gap-1.5 py-px text-muted-foreground">
-        <TerminalSquare className="h-3.5 w-3.5 shrink-0 opacity-70" />
-        <span className="truncate text-xs font-medium">{item.title}</span>
-        <span className="shrink-0 text-xs text-muted-foreground/70">
-          {item.sections.length} section{item.sections.length === 1 ? "" : "s"}
-        </span>
-        <TranscriptTimestamp timestamp={item.timestamp} />
-        <ChevronDown className="h-4 w-4 shrink-0 transition-transform group-open:rotate-180" />
-      </summary>
-      <div className="space-y-3 py-2 pl-5">
-        {item.sections.map((section) => (
-          <details
-            className="group/section"
-            key={`${section.title}:${section.body.slice(0, 48)}`}
-          >
-            <summary className="inline-flex max-w-full cursor-pointer list-none items-center gap-1.5 text-xs font-medium text-foreground/80">
-              <span className="truncate">{section.title}</span>
-              <ChevronDown className="h-4 w-4 shrink-0 text-muted-foreground transition-transform group-open/section:rotate-180" />
-            </summary>
-            <pre className="mt-2 max-h-56 overflow-auto whitespace-pre-wrap wrap-break-word rounded-md bg-muted/50 px-3 py-2 font-mono text-xs leading-5 text-muted-foreground">
-              {section.body.trim() || "No metadata."}
-            </pre>
-          </details>
-        ))}
-      </div>
-    </details>
-  );
-}
-
-function LifecycleItem({
-  item,
-}: {
-  item: Extract<TranscriptItem, { type: "lifecycle" }>;
-}) {
-  const isError = item.title.toLowerCase().includes("error");
-  return (
-    <div
-      className={cn(
-        "flex items-center justify-start gap-1.5 rounded-md px-2 py-1.5 text-left text-xs",
-        isError
-          ? "border border-destructive/20 bg-destructive/5 text-destructive"
-          : "text-muted-foreground/80",
-      )}
-      data-testid="transcript-lifecycle-item"
-    >
-      {isError ? (
-        <AlertCircle className="h-3.5 w-3.5 shrink-0" />
-      ) : (
-        <CircleDot className="h-3 w-3 shrink-0 opacity-50" />
-      )}
-      <span className="font-medium">{item.title}</span>
-      {item.text ? <span className="opacity-80">· {item.text}</span> : null}
-      <TranscriptTimestamp timestamp={item.timestamp} />
-    </div>
-  );
-}
-
-const fullDateTimeFormat = new Intl.DateTimeFormat(undefined, {
-  weekday: "long",
-  year: "numeric",
-  month: "long",
-  day: "numeric",
-  hour: "numeric",
-  minute: "2-digit",
-  second: "2-digit",
-});
-
-function TranscriptTimestamp({ timestamp }: { timestamp: string }) {
-  const formatted = formatTranscriptTime(timestamp);
-  if (!formatted) return null;
-  const date = new Date(timestamp);
-  const fullDateTime = Number.isNaN(date.getTime())
-    ? timestamp
-    : fullDateTimeFormat.format(date);
-  return (
-    <Tooltip>
-      <TooltipTrigger asChild>
-        <span className="shrink-0 cursor-default text-xs text-muted-foreground/60">
-          {formatted}
-        </span>
-      </TooltipTrigger>
-      <TooltipContent side="top">{fullDateTime}</TooltipContent>
-    </Tooltip>
-  );
-}
