+++ desktop/src/features/agents/ui/activityRenderClasses/ErrorActivity.tsx
@@ -1,0 +1,6 @@
+import { LifecycleActivity } from "./LifecycleActivity";
+import type { ActivityRenderClassItemProps } from "./types";
+
+export function ErrorActivity(props: ActivityRenderClassItemProps) {
+  return <LifecycleActivity {...props} />;
+}
