+++ desktop/src/features/agents/ui/activityRenderClasses/FileEditActivity.tsx
@@ -1,0 +1,6 @@
+import { ToolActivity } from "./ToolActivity";
+import type { ActivityRenderClassItemProps } from "./types";
+
+export function FileEditActivity(props: ActivityRenderClassItemProps) {
+  return <ToolActivity {...props} />;
+}
