+++ desktop/src/features/agents/ui/activityRenderClasses/GenericActivity.tsx
@@ -1,0 +1,6 @@
+import { ToolActivity } from "./ToolActivity";
+import type { ActivityRenderClassItemProps } from "./types";
+
+export function GenericActivity(props: ActivityRenderClassItemProps) {
+  return <ToolActivity {...props} />;
+}
