+++ desktop/src/features/agents/ui/activityRenderClasses/PermissionActivity.tsx
@@ -1,0 +1,6 @@
+import { LifecycleActivity } from "./LifecycleActivity";
+import type { ActivityRenderClassItemProps } from "./types";
+
+export function PermissionActivity(props: ActivityRenderClassItemProps) {
+  return <LifecycleActivity {...props} />;
+}
