+++ desktop/src/features/agents/ui/activityRenderClasses/RelayOpActivity.tsx
@@ -1,0 +1,6 @@
+import { ToolActivity } from "./ToolActivity";
+import type { ActivityRenderClassItemProps } from "./types";
+
+export function RelayOpActivity(props: ActivityRenderClassItemProps) {
+  return <ToolActivity {...props} />;
+}
