+++ desktop/src/features/agents/ui/activityRenderClasses/ShellActivity.tsx
@@ -1,0 +1,6 @@
+import { ToolActivity } from "./ToolActivity";
+import type { ActivityRenderClassItemProps } from "./types";
+
+export function ShellActivity(props: ActivityRenderClassItemProps) {
+  return <ToolActivity {...props} />;
+}
