+++ desktop/src/features/agents/ui/activityRenderClasses/StatusActivity.tsx
@@ -1,0 +1,6 @@
+import { LifecycleActivity } from "./LifecycleActivity";
+import type { ActivityRenderClassItemProps } from "./types";
+
+export function StatusActivity(props: ActivityRenderClassItemProps) {
+  return <LifecycleActivity {...props} />;
+}
