+++ desktop/src/features/agents/ui/activityRenderClasses/SuppressedActivity.tsx
@@ -1,0 +1,6 @@
+import { ToolActivity } from "./ToolActivity";
+import type { ActivityRenderClassItemProps } from "./types";
+
+export function SuppressedActivity(props: ActivityRenderClassItemProps) {
+  return <ToolActivity {...props} />;
+}
