+++ desktop/src/features/agents/ui/activityRenderClasses/ToolActivity.tsx
@@ -1,0 +1,10 @@
+import { ToolItem } from "../AgentSessionToolItem";
+import type { ActivityRenderClassItemProps } from "./types";
+
+export function ToolActivity({ item }: ActivityRenderClassItemProps) {
+  if (item.type !== "tool") {
+    return null;
+  }
+
+  return <ToolItem item={item} />;
+}
