+++ desktop/src/features/agents/ui/activityRenderClasses/TranscriptTimestamp.tsx
@@ -1,0 +1,31 @@
+import { Tooltip, TooltipContent, TooltipTrigger } from "@/shared/ui/tooltip";
+import { formatTranscriptTime } from "../agentSessionUtils";
+
+const fullDateTimeFormat = new Intl.DateTimeFormat(undefined, {
+  weekday: "long",
+  year: "numeric",
+  month: "long",
+  day: "numeric",
+  hour: "numeric",
+  minute: "2-digit",
+  second: "2-digit",
+});
+
+export function TranscriptTimestamp({ timestamp }: { timestamp: string }) {
+  const formatted = formatTranscriptTime(timestamp);
+  if (!formatted) return null;
+  const date = new Date(timestamp);
+  const fullDateTime = Number.isNaN(date.getTime())
+    ? timestamp
+    : fullDateTimeFormat.format(date);
+  return (
+    <Tooltip>
+      <TooltipTrigger asChild>
+        <span className="shrink-0 cursor-default text-xs text-muted-foreground/60">
+          {formatted}
+        </span>
+      </TooltipTrigger>
+      <TooltipContent side="top">{fullDateTime}</TooltipContent>
+    </Tooltip>
+  );
+}
